+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.0.2] - 2024-06-27
 ### Added
 - add tests with drupal 10.3
 - add configuration settings `timesup.settings.resolvers` to enable or disable resolvers.
@@ -66,7 +68,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - init module
 
-[Unreleased]: https://github.com/antistatique/drupal-timesup/compare/2.0.1...HEAD
+[Unreleased]: https://github.com/antistatique/drupal-timesup/compare/2.0.2...HEAD
+[2.0.2]: https://github.com/antistatique/drupal-timesup/compare/2.0.1...2.0.2
 [2.0.1]: https://github.com/antistatique/drupal-timesup/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/antistatique/drupal-timesup/compare/8.x-1.1...2.0.0
 [1.1.0]: https://github.com/antistatique/drupal-timesup/compare/8.x-1.0...8.x-1.1